# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": "Documentação da API do backend"
  },
  "servers": [
    {
      "url": "https://tematicoapi.zapchatbr.com/",
      "description": "API Temático"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "summary": "Criar usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado"
          },
          "409": {
            "description": "Usuário ou email já existe"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Obter usuário pelo ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "summary": "Atualizar usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "summary": "Desativar usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário desativado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/restaurantes": {
      "get": {
        "summary": "Listar restaurantes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de restaurantes"
          }
        },
        "tags": [
          "Restaurantes"
        ]
      },
      "post": {
        "summary": "Criar restaurante",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "capacidade": {
                    "type": "integer"
                  },
                  "descricao": {
                    "type": "string"
                  }
                },
                "required": [
                  "nome",
                  "capacidade"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Restaurante criado"
          }
        },
        "tags": [
          "Restaurantes"
        ]
      }
    },
    "/restaurantes/{id}": {
      "get": {
        "summary": "Obter restaurante pelo ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurante encontrado"
          },
          "404": {
            "description": "Restaurante não encontrado"
          }
        },
        "tags": [
          "Restaurantes"
        ]
      },
      "put": {
        "summary": "Atualizar restaurante",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "capacidade": {
                    "type": "integer"
                  },
                  "descricao": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restaurante atualizado"
          },
          "404": {
            "description": "Restaurante não encontrado"
          }
        },
        "tags": [
          "Restaurantes"
        ]
      },
      "delete": {
        "summary": "Deletar restaurante",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurante deletado"
          },
          "404": {
            "description": "Restaurante não encontrado"
          }
        },
        "tags": [
          "Restaurantes"
        ]
      }
    },
...
  }
}